% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{comparison}
\alias{comparison}
\title{Create a \code{comparison} object}
\usage{
comparison(...)
}
\arguments{
\item{...}{Multiple \code{scenario}s or a list of \code{scenario} objects.}
}
\value{
A \code{comparison} object.
}
\description{
The \code{comparison} class is intended to store \code{scenario} data and
to quickly run comparisons against a user-specified baseline.
}
