% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_abm_eg}
\alias{covid_abm_eg}
\title{Example Dataset}
\format{
\subsection{\code{covid_abm_eg}}{

A data.table with ~1.1M rows and 13 columns.
}
}
\source{
\url{https://medrxiv.org}
}
\usage{
covid_abm_eg
}
\description{
A subset of data from a COVID modeling simulation study
}
\keyword{datasets}
