% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{is.scenario}
\alias{is.scenario}
\title{Check whether an object is a 'scenario'}
\usage{
is.scenario(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
A logical indicating whether the object inherits from the class
'scenario'.
}
\description{
Check whether an object is a 'scenario'
}
\examples{
# prepare two scenarios of the final size of an epidemic
pandemic_flu <- scenario(
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(r0 = 1.5),
  replicates = 1L
)

is.scenario(pandemic_flu)
}
