% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{is_comparison}
\alias{is_comparison}
\title{Check whether an object is a \code{comparison}}
\usage{
is_comparison(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
A logical indicating whether the object inherits from the class
\code{comparison}.
}
\description{
Check whether an object is a \code{comparison}
}
