% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenient_parameters.R
\name{make_parameters_finalsize_UK}
\alias{make_parameters_finalsize_UK}
\title{Make useful default arguments to \code{finalsize::final_size()}}
\usage{
make_parameters_finalsize_UK(r0 = 1.5)
}
\arguments{
\item{r0}{The basic reproductive number of an infection. Default is 1.5 for
pandemic influenza.}
}
\value{
A list of arguments to \code{\link[finalsize:final_size]{finalsize::final_size()}}.
}
\description{
Helper function to make useful arguments to \code{\link[finalsize:final_size]{finalsize::final_size()}},
with the possibility to specify \eqn{R_0}, which is set to 1.5 by default.
Draws on the default U.K. POLYMOD social contacts dataset included in
\code{{finalsize}}, with three age groups: 0 -- 20, 20 -- 40, and > 40.
Assumes a single susceptibility group, with full susceptibility.
Selects the Newton solver.
}
\examples{
pandemic_flu_args <- make_parameters_finalsize_UK(r0 = 1.5)
covid_args <- make_parameters_finalsize_UK(r0 = 5.0)
}
