% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{new_comparison}
\alias{new_comparison}
\title{Constructor for the 'comparison' class}
\usage{
new_comparison(data, baseline = 1, match_variables, comparison_variables)
}
\arguments{
\item{data}{A list of 'scenario' objects.}

\item{baseline}{A string or integer for the element of the list of 'scenario'
objects which indicates which should be considered the 'baseline' outcome,
against which other outcomes are compared.}

\item{match_variables}{The variables in the 'scenario' outputs on which to
match the scenarios and check whether they are comparable.}

\item{comparison_variables}{The variables in the 'scenario' outputs to
compare against the 'baseline' scenario.}
}
\value{
A 'comparison' object
}
\description{
Create a 'comparison' object after input checks.
}
\keyword{internal}
