% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{new_scenario}
\alias{new_scenario}
\title{Construct a \code{scenario} object}
\usage{
new_scenario(
  data,
  name = NA_character_,
  model = NA_character_,
  parameters = list(),
  extra_info = list()
)
}
\arguments{
\item{data}{Any object that can be sensibly coerced to a \code{data.table}, representing
the epidemiological model data.}

\item{name}{String of the senario name. Defaults to \code{NA} if not provided.}

\item{model}{String of the model name. Defaults to \code{NA} if not provided.}

\item{parameters}{Named list of the model parameters. Only atomic list
elements are allowed.}

\item{extra_info}{Extra model information that may be useful for matching
models. Only atomic list elements are supported.}
}
\value{
A \code{scenario} object.
}
\description{
Create a scenario object with input checks.
}
\keyword{internal}
