% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scenario.R
\name{run_scenario.comparison}
\alias{run_scenario.comparison}
\title{Run scenarios contained in a comparison}
\usage{
\method{run_scenario}{comparison}(x)
}
\arguments{
\item{x}{A \code{scenario} object.}
}
\value{
A \code{comparison} with the attributes of the original object passed to
the function, but with the \code{data} list containing \code{scenario} objects with
replicate outcomes.
}
\description{
Run the epidemic model scenarios specified as \code{scenario} objects
in a \code{comparison} object. Calls \code{\link[=run_scenario.scenario]{run_scenario.scenario()}} on each object in
the \code{data} list of the comparison object.
}
