% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{run_scenario}
\alias{run_scenario}
\title{Run a scenario}
\usage{
run_scenario(x)
}
\arguments{
\item{x}{A \code{scenario} object.}
}
\value{
The original \code{scenario} object with the \code{data} field populated with
simulation output. This object must be assigned.
}
\description{
Run an epidemic model scenario using the function stored in
\code{model_function} with the arguments in \code{parameters}. Runs as many replicates
of the function as specified in \code{replicates}. The simulation output data are
stored under \code{data}, and the \code{data_available} tag is updated to \code{TRUE}.
}
\examples{
# prepare a scenario
scenario_pandemic_flu <- scenario(
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(), # using helper function
  replicates = 1L
)

# print to check that data are not prepared
scenario_pandemic_flu

# generate scenario data
scenario_pandemic_flu <- run_scenario(scenario_pandemic_flu)

# print to check that data are prepared
scenario_pandemic_flu
}
