% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scenario.R
\name{run_scenario.scenario}
\alias{run_scenario.scenario}
\title{Run a scenario}
\usage{
\method{run_scenario}{scenario}(x)
}
\arguments{
\item{x}{A \code{scenario} object.}
}
\value{
The original \code{scenario} object with the \code{data} field populated with
simulation output. This object must be assigned.
}
\description{
Run an epidemic model scenario using the function stored in
\code{model_function} with the arguments in \code{parameters}. Runs as many replicates
of the function as specified in \code{replicates}. The simulation output data are
stored under \code{data}.
}
