% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sce_add_info}
\alias{sce_add_info}
\title{Add extra information to a scenario}
\usage{
sce_add_info(x, info)
}
\arguments{
\item{x}{A \code{scenario} object.}

\item{info}{A named list of information to be added to the \code{extra_info} list
of the scenario object \code{x}.}
}
\value{
The scenario \code{x} with the extra information added.
}
\description{
Add extra information to a scenario
}
\examples{
# get some parameters for a `finalsize` run
parameters <- make_parameters_finalsize_UK(r0 = 1.5)
extra_info <- list(
  age_groups = rownames(parameters$contact_matrix)
)
x <- scenario(
  model_function = "finalsize::final_size",
  parameters = parameters
)

sce_add_info(x, extra_info)

}
