% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sce_aggregate_outcomes}
\alias{sce_aggregate_outcomes}
\title{Aggregate scenario outcomes across replicates}
\usage{
sce_aggregate_outcomes(
  x,
  grouping_variables,
  measure_variables,
  summary_functions = c("mean", "sd")
)
}
\arguments{
\item{x}{A scenario object with data prepared.}

\item{grouping_variables}{The variables that should be used to group the
outcomes of interest. Examples include demographic and susceptibility groups.}

\item{measure_variables}{The outcomes of interest which are summarised over
scenario replicates, and by all variables in the \code{grouping_variables}.
Examples include \code{p_infected} from the output of \code{\link[finalsize:final_size]{finalsize::final_size()}}.}

\item{summary_functions}{The summary function names to apply to the measure
variables, passed as strings, i.e., "mean" rather than simply \code{mean}.}
}
\value{
A data.table with the outcomes of interest (measure variables)
summarised using the summary functions, by each grouping variable.
}
\description{
Aggregate scenario outcomes across replicates
}
\examples{
# create a scenario
scenario_pandemic_flu <- scenario(
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(),
  replicates = 3 # note extra replicates
)

# run scenario
scenario_pandemic_flu <- run_scenario(scenario_pandemic_flu)

# peek at outcome to see column names
sce_peek_outcomes(scenario_pandemic_flu)

# aggregate outcome by demographic group
sce_aggregate_outcomes(
  x = scenario_pandemic_flu,
  grouping_variables = c("demo_grp"),
  measure_variables = c("p_infected"),
  summary_functions = c("mean", "min", "max")
)
}
