% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparing_scenarios.R
\name{sce_are_comparable}
\alias{sce_are_comparable}
\title{Check whether a pair of scenarios is comparable}
\usage{
sce_are_comparable(baseline, compare, match_variables, comparison_variables)
}
\arguments{
\item{baseline}{A 'scenario' object.}

\item{compare}{A second 'scenario' object.}

\item{match_variables}{A character string of scenario parameter names that is
used to check whether the two scenarios have identical parameters.}

\item{comparison_variables}{A character string of column names expected in
the scenarios' outcome data. This is used to check whether the scenarios both
have the required columns in their data.}
}
\value{
A logical value indicating whether a pair of scenarios is comparable.
}
\description{
Check whether a pair of scenarios is comparable
}
\examples{
# prepare two scenarios of the final size of an epidemic
pandemic_flu <- scenario(
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(r0 = 1.5),
  replicates = 1L
)

covid19 <- scenario(
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(r0 = 5.0),
  replicates = 1L
)

# run scenarios to generate data
pandemic_flu <- run_scenario(pandemic_flu)
covid19 <- run_scenario(covid19)

# check whether scenarios are comparable
sce_are_comparable(
  baseline = pandemic_flu,
  compare = covid19,
  match_variables = "demography_vector",
  comparison_variables = "p_infected"
)
}
