% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sce_drop_data.comparison}
\alias{sce_drop_data.comparison}
\title{Drop outcome data from a \code{comparison} object}
\usage{
\method{sce_drop_data}{comparison}(x)
}
\arguments{
\item{x}{A \code{comparison} object.}
}
\value{
A \code{comparison} object where the \code{data} field is populated with
\code{scenario} objects from which the outcome data, if originally present,
has been dropped.
}
\description{
Drop outcome data from a \code{comparison} object
}
