% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sce_drop_data}
\alias{sce_drop_data}
\title{Drop outcome data from a \code{scenario} or \code{comparison} object}
\usage{
sce_drop_data(x)
}
\arguments{
\item{x}{A \code{scenario} or \code{comparison} object.}
}
\value{
A \code{scenario} or \code{comparison} object where the \code{data} field is, for
\code{scenario} objects, an empty list of the same length as the number of
scenario replicates, and for \code{comparison} objects, a list of \code{scenario}
objects with the data removed.
}
\description{
Drop outcome data from a \code{scenario} or \code{comparison} object
}
\examples{
pandemic_flu <- scenario(
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(r0 = 1.5),
  extra_info = list(country = "UK", pathogen = "flu")
)

# run scenarios to generate data
pandemic_flu <- run_scenario(pandemic_flu)

# drop data
sce_drop_data(pandemic_flu)
}
