% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sce_drop_data.scenario}
\alias{sce_drop_data.scenario}
\title{Drop outcome data from a \code{scenario} object}
\usage{
\method{sce_drop_data}{scenario}(x)
}
\arguments{
\item{x}{A \code{scenario} object.}
}
\value{
A \code{scenario} object where the \code{data} field is an empty list of the
same length as the number of scenario replicates.
}
\description{
Drop outcome data from a \code{scenario} object
}
