% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparing_scenarios.R
\name{sce_filter_comparable}
\alias{sce_filter_comparable}
\title{Filter for comparable scenarios}
\usage{
sce_filter_comparable(x, match_variables, comparison_variables)
}
\arguments{
\item{x}{A 'comparison' object comprised of 'scenario' objects, each of which
must have data prepared.}

\item{match_variables}{A character string of scenario parameter names that is
used to check whether the two scenarios have identical parameters.}

\item{comparison_variables}{A character string of column names expected in
the scenarios' outcome data. This is used to check whether the scenarios both
have the required columns in their data.}
}
\value{
A 'comparison' object retaining only those 'scenario' objects in the
data list that are comparable with the baseline scenario.
}
\description{
Filter for comparable scenarios
}
\examples{
pandemic_flu <- scenario(
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(r0 = 1.5),
  extra_info = list(country = "UK", pathogen = "flu")
)

covid19 <- scenario(
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(r0 = 5.0),
  extra_info = list(country = "UK", pathogen = "SARS-CoV-2")
)

# run scenarios to generate data
pandemic_flu <- run_scenario(pandemic_flu)
covid19 <- run_scenario(covid19)

# check whether scenarios are comparable
outbreak_comparison <- comparison(
  pandemic_flu = pandemic_flu, covid19 = covid19,
  baseline = "pandemic_flu"
)

# filter on pathogen
sce_filter_comparable(
  outbreak_comparison,
  match_variables = "pathogen",
  comparison_variables = "p_infected"
)
}
