% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{sce_filter_comparison}
\alias{sce_filter_comparison}
\title{Filter \code{comparison} data by key-values}
\usage{
sce_filter_comparison(x, matching_variables)
}
\arguments{
\item{x}{A \code{comparison} object.}

\item{matching_variables}{Named list of the key-value pairs by which to
filter the \code{comparison} data. The list-element names are used to key the data
and the list-element values are used to filter the data.}
}
\value{
A \code{comparison} object.
}
\description{
Filter \code{comparison} data by key-values
}
