% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export.R
\name{sce_from_json}
\alias{sce_from_json}
\title{Read scenario from JSON}
\usage{
sce_from_json(file)
}
\arguments{
\item{file}{A JSON file}
}
\value{
A \code{scenario} object.
}
\description{
Read scenario from JSON
}
\examples{
scenario_pandemic_flu <- scenario(
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(),
  replicates = 1
)
# create temporary filepath
tmpfile <- tempfile(fileext = ".json")
sce_to_json(scenario_pandemic_flu, file = tmpfile)
sce_from_json(tmpfile)
}
