% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcomes.R
\name{sce_get_outcomes.comparison}
\alias{sce_get_outcomes.comparison}
\title{Get scenario outcomes from a comparison object}
\usage{
\method{sce_get_outcomes}{comparison}(x)
}
\arguments{
\item{x}{A \code{comparison} object with data prepared. Check for whether data has
been prepared using \code{\link[=sce_has_data]{sce_has_data()}}.}
}
\value{
A single data.table holding the outcomes of each replicate of each
scenario specified in \code{x}. Scenarios are identified either by their names if
any are provided in the \code{data} list in \code{x}, or by a simple synthetic
identifier such as 'scenario_01'.
}
\description{
Get scenario outcomes from a comparison object
}
