% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcomes.R
\name{sce_get_outcomes}
\alias{sce_get_outcomes}
\title{Get scenario outcomes}
\usage{
sce_get_outcomes(x)
}
\arguments{
\item{x}{A \code{scenario} or \code{comparison} object with the scenario outcome data
prepared. Check for whether data has been prepared using \code{\link[=sce_has_data]{sce_has_data()}}.}
}
\value{
A single \code{data.table} holding the output of all replicates of the
scenario. If \code{x} is a \code{comparison} object, the \code{data.table} holds the
outcomes of each replicate of each scenario specified therein. Scenarios in
this case are identified either by their names if any, or by a simple
synthetic identifier such as 'scenario_01'.
}
\description{
Get scenario outcomes
}
