% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcomes.R
\name{sce_get_outcomes.scenario}
\alias{sce_get_outcomes.scenario}
\title{Get scenario outcomes from a scenario object}
\usage{
\method{sce_get_outcomes}{scenario}(x)
}
\arguments{
\item{x}{A scenario object with data prepared. Check for whether data has
been prepared using \code{\link[=sce_has_data]{sce_has_data()}}.}
}
\value{
A single data.table holding the output of all replicates of the
scenario. Contains the \code{replicate} column to help differentiate data from
each replicate.
}
\description{
Get scenario outcomes from a scenario object
}
\examples{
# create a scenario
scenario_pandemic_flu <- scenario(
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(),
  replicates = 3 # note extra replicates
)

# run scenario
scenario_pandemic_flu <- run_scenario(scenario_pandemic_flu)

# get outcomes
sce_get_outcomes(scenario_pandemic_flu)
}
