% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sce_get_parameters}
\alias{sce_get_parameters}
\title{Print scenario parameters}
\usage{
sce_get_parameters(x, which = NULL)
}
\arguments{
\item{x}{A \code{\link[=scenario]{scenario()}} object.}

\item{which}{Which parameters to print.}
}
\value{
Nothing. Prints a list of parameters to screen.
}
\description{
Print scenario parameters
}
\examples{
# create a scenario
scenario_pandemic_flu <- scenario(
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(),
  replicates = 1
)

# get all parameters
sce_get_parameters(scenario_pandemic_flu)

# get only some parameters
sce_get_parameters(scenario_pandemic_flu, which = c("r0", "solver"))
}
