% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_helpers.R
\name{sce_get_scenario_names}
\alias{sce_get_scenario_names}
\title{Get scenario names from a comparison object}
\usage{
sce_get_scenario_names(x)
}
\arguments{
\item{x}{A comparison object.}
}
\value{
A vector of scenario names.
}
\description{
Get scenario names from a comparison object
}
\examples{
# create some scenarios
pandemic_flu <- scenario(
  name = "pandemic_flu",
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(r0 = 1.5),
  replicates = 1L
)

covid19 <- scenario(
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(r0 = 5.0),
  replicates = 1L
)

# create comparison, passing prepared scenarios with some names
outbreak_comparison <- comparison(
  pandemic_flu, covid19,
  baseline = "pandemic_flu"
)
}
