% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparing_scenarios.R
\name{sce_has_comparison_variables}
\alias{sce_has_comparison_variables}
\title{Check for variables in scenario outcomes}
\usage{
sce_has_comparison_variables(x, variables)
}
\arguments{
\item{x}{A \code{scenario} object.}

\item{variables}{A character vector of variables that are expected to be
found in the \code{data} list of \code{x}.}
}
\value{
A logical indicating whether the variables correspond to column names
in \code{x$data}.
}
\description{
Check for variables in scenario outcomes
}
\keyword{internal}
