% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sce_has_data}
\alias{sce_has_data}
\title{Check for scenario data}
\usage{
sce_has_data(x)
}
\arguments{
\item{x}{A \code{scenario} object.}
}
\value{
A boolean, whether the simulation object has data.
}
\description{
Check for scenario data
}
\examples{
# create a scenario
scenario_pandemic_flu <- scenario(
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(),
  replicates = 1
)

sce_has_data(scenario_pandemic_flu)
}
