% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sce_has_data.scenario}
\alias{sce_has_data.scenario}
\title{Check for scenario data}
\usage{
\method{sce_has_data}{scenario}(x)
}
\arguments{
\item{x}{A 'scenario' object.}
}
\value{
A boolean, whether the simulation object has data.
}
\description{
Check for scenario data
}
