% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparing_scenarios.R
\name{sce_has_match_variables}
\alias{sce_has_match_variables}
\title{Check for scenario matching variables}
\usage{
sce_has_match_variables(x, variables)
}
\arguments{
\item{x}{A \code{scenario} object.}

\item{variables}{A character vector of \code{scenario} parameter names or extra
information names for which to search in \code{x}.}
}
\value{
A logical indicating whether the variables correspond to parameter
names in \code{x}.
}
\description{
Check for scenario matching variables
}
\keyword{internal}
