% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sce_peek_outcomes}
\alias{sce_peek_outcomes}
\title{Get scenario outcome names}
\usage{
sce_peek_outcomes(x, view_rows = FALSE)
}
\arguments{
\item{x}{A scenario object with data prepared. Check for whether data has
been prepared using \code{\link[=sce_has_data]{sce_has_data()}}.}

\item{view_rows}{Whether to return the first few rows of the first replicate
outcome, using \code{\link[=head]{head()}}.}
}
\value{
Prints the scenario outcome data's column names and types to screen,
or the \code{\link[=head]{head()}} of the first outcome replicate if \code{view_rows = TRUE}.
}
\description{
Function to quickly view the names and types of columns in the \code{scenario}
outcome data. Operates on the first replicate of each scenario run, and
assumes that the outcome data are data.frames.
}
\examples{
# create a scenario
scenario_pandemic_flu <- scenario(
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(),
  replicates = 1
)
run_scenario(scenario_pandemic_flu)

# for column names and types
sce_peek_outcomes(scenario_pandemic_flu)
# for data.frame head
sce_peek_outcomes(scenario_pandemic_flu, view_rows = FALSE)
}
