% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export.R
\name{sce_to_json}
\alias{sce_to_json}
\title{Write scenario to JSON}
\usage{
sce_to_json(x, file)
}
\arguments{
\item{x}{A \code{scenario} object.}

\item{file}{The file to which to write the JSON.}
}
\value{
Writes the \code{scenario} object to JSON.
}
\description{
Write scenario to JSON
}
\examples{
scenario_pandemic_flu <- scenario(
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(),
  replicates = 1
)

sce_to_json(scenario_pandemic_flu, file = tempfile())
}
