% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{scenario}
\alias{scenario}
\title{Create a \code{scenario} object}
\usage{
scenario(
  data = data.table::data.table(),
  name = NA_character_,
  model = NA_character_,
  parameters = list(),
  extra_info = list()
)
}
\arguments{
\item{data}{Any object that can be sensibly coerced to a \code{data.table}, representing
the epidemiological model data.}

\item{name}{String of the senario name. Defaults to \code{NA} if not provided.}

\item{model}{String of the model name. Defaults to \code{NA} if not provided.}

\item{parameters}{Named list of the model parameters. Only atomic list
elements are allowed.}

\item{extra_info}{Extra model information that may be useful for matching
models. Only atomic list elements are supported.}
}
\value{
A \code{scenario} object
}
\description{
The \code{scenario} class is intended to store the outcomes of an
epidemic simulation.
}
\examples{
# prepare arguments to `finalsize::final_size()`
# using the included convenience function
pandemic_flu_args <- make_parameters_finalsize_UK(r0 = 1.5)

# prepare extra information on age group limits
age_groups <- rownames(pandemic_flu_args$contact_matrix)

data <- do.call(
  finalsize::final_size,
  pandemic_flu_args
)

scenario(
  data = data,
name = "final_size_UK",
  model = "final_size",
  parameters = list(
 R0 = 1.5
),
  extra_info = list(
    n_age_groups = 3
  )
)
}
