% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios-package.R
\docType{package}
\name{scenarios-package}
\alias{scenarios}
\alias{scenarios-package}
\title{scenarios: Convenience Methods for Scenario Analysis}
\description{
`{scenarios}` simplifies scenario analysis by managing the bookkeeping work for you.
}
\author{
\strong{Maintainer}: Carl Pearson \email{carl.ab.pearson@gmail.com} (\href{https://orcid.org/0000-0003-0701-7860}{ORCID})

}
\keyword{internal}
