% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{scn_analyze}
\alias{scn_analyze}
\title{Execute Scenario Analysis}
\usage{
scn_analyze(
  outcomes,
  plan,
  compare,
  summarize,
  inputs,
  group,
  joinon = setdiff(names(outcomes), inputs)
)
}
\arguments{
\item{outcomes_dt}{a set of outputs, as produced by \code{\link[=scn_outcomes]{scn_outcomes()}}}
}
\description{
Conduct a scenario analysis according to a particular comparison plan, for
a set of given comparisons, optionally summarized.
}
