% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{scn_ceff}
\alias{scn_ceff}
\title{Cumulative Effectiveness Comparator}
\usage{
scn_ceff(colname = "value")
}
\arguments{
\item{colname}{a string, the column name in the output that will be compared}
}
\value{
a expression, which will be evaluated in the context of the output
}
\description{
Convenience function to create a cumulative effectiveness comparator for use
in \code{\link[=scn_analyze]{scn_analyze()}} for the \code{compare} argument.
}
\examples{
# by default, assume a `melt.data.table`-like result, with `value` as the
# column
scn_ceff()

# but if in wide-format, can instead give the column names
scn_ceff("infections")
scn_ceff("deaths")
}
