% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{scn_convert}
\alias{scn_convert}
\title{Convert Data to be Keyed by Scenario

This method takes data which has existing scenario columns, and returns that
data converted to be in terms of the scenario keys.}
\usage{
scn_convert(data, scen_dt, keep = setdiff(names(data), names(scen_dt)))
}
\arguments{
\item{data}{a data.table, such as might be provided to \code{\link[=scn_extract]{scn_extract()}}}

\item{scen_dt}{a data.table, meeting the return criteria of \code{\link[=scn_extract]{scn_extract()}}}

\item{keep}{the columns to keep; by default, all of the columns that columns
that are \emph{not} defining scenarios.}
}
\value{
a data.table, with columns \code{keep} + \code{scen_id}
}
\description{
Convert Data to be Keyed by Scenario

This method takes data which has existing scenario columns, and returns that
data converted to be in terms of the scenario keys.
}
