% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{scn_extract}
\alias{scn_extract}
\title{Extract Scenario Table}
\usage{
scn_extract(
  data,
  exclude = data.table::key(data),
  orderby = setdiff(names(data), exclude)
)
}
\arguments{
\item{exclude}{column names to exclude. defaults to \code{\link[data.table:setkey]{data.table::key()}} of \code{dt}}

\item{orderby}{for the output, what column priority to order on, which
determines the scenario labeling order.}

\item{dt}{a \code{\link[=data.table]{data.table()}} of parameter / settings}
}
\value{
a \code{\link[=data.table]{data.table()}} with an \code{id} column and where all other columns form
the unique combinations of the non-excluded columns in dt.
}
\description{
Extract scenario table from a full table of parameters, settings, etc.
}
\details{
This method extracts the unique scenario set from a parameter table. It is
easiest to use on a data.table, that already has a key, where that key
is like a primary key in a relational data sense.

Note: if your data already have something like a scenario table, you don't
need to use this method, but you should make sure that scenario table
conforms to the return value for this function: an \code{id} column, and all the
unique combinations of feature columns.
}
\examples{
library(scenarios)
data(covid_abm_eg)
covid_abm_eg |> scn_extract(c("realization", "date", "inf", "deaths"))

}
