% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{scn_grid}
\alias{scn_grid}
\title{Create an Analysis Grid}
\usage{
scn_grid(data, plan, match_by = key(data))
}
\description{
Creates a tabular view of outputs based on a comparison plan + matched output
features
}
