% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{scn_outcomes}
\alias{scn_outcomes}
\title{title}
\usage{
scn_outcomes(data, scen_dt)
}
\value{
a reduced view of \code{data}, replacing all scenario information with a
\code{scen_id} column.
}
\description{
title
}
