% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{scn_plan}
\alias{scn_plan}
\title{Create a Scenario Analysis Plan}
\usage{
scn_plan(scen_dt, baseline, ignore = c())
}
\arguments{
\item{scen_dt}{a \code{\link[data.table:data.table]{data.table::data.table()}} as returned by \code{\link[=scn_extract]{scn_extract()}} and
passing \code{\link[=scn_check]{scn_check()}}}

\item{baseline}{creates a plan where features match, except those named
by this argument. The scenarios with the corresponding values for these names
become the reference values}

\item{ignore}{other columns that can be ignored when matching \code{baseline}. The
\code{scen_id} column is always automatically ignored}
}
\value{
a data.table, with columns \code{comp}(arison) and \code{base}(line), both
integer id columns, corresponding to values from \code{scen_dt}
}
\description{
A short description...
}
