% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregators.R
\name{scn_quantile}
\alias{scn_quantile}
\title{Quantile Summarizer}
\usage{
scn_quantile(probs = c(lo = 0.025, md = 0.5, hi = 0.975))
}
\arguments{
\item{probs}{like the \code{probs} argument to \code{\link[stats:quantile]{stats::quantile()}}, but supports names
The names will be used as the column names.}
}
\value{
a function, suitable for application to long-format data in grouped chunks
}
\description{
A summarizer for use in \code{\link[=scn_analyze]{scn_analyze()}} for \code{summary} argument. Creates quantiles
}
