% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{scn_validate}
\alias{scn_validate}
\title{Validate a Potential Scenario Table}
\usage{
scn_validate(scen_dt)
}
\arguments{
\item{scen_dt}{a data.table, such as would be returned by \code{\link[=scn_extract]{scn_extract()}}.
If the \code{scen_dt} does conform to the requirements for a scenario table,
throws an informative error.}
}
\value{
for convenience, returns \code{scen_dt} unmodified (if validated)
}
\description{
Validate a Potential Scenario Table
}
