% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{scn_view}
\alias{scn_view}
\title{Extract Baseline or Comparison Views}
\usage{
scn_view(data, plan, mode = c("baseline", "comparison"))
}
\arguments{
\item{data}{a data.table, like returned from \code{\link[=scn_convert]{scn_convert()}} - i.e. having
a \code{scen_id} column, other outcome keying columns, and then outcome values}

\item{plan}{a data.table, like returned from \code{\link[=scn_plan]{scn_plan()}} - i.e. having
\code{comp} and \code{base} columns that correspond to scenario ids.}

\item{mode}{a string, either "baseline" or "comparison", indicating which
view to return}
}
\value{
a filtered copy of \code{data}, with either the baseline scenario(s) or
the comparison ones, depending on \code{mode} argument.
}
\description{
\code{{scenarios}} is primarily for preparing matched comparative analyses, but
you may also need to extract the uncompared views. The function let's you
use your analysis plan from \code{\link[=scn_plan]{scn_plan()}} to extract your baseline or
comparison scenarios.
}
