% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{validate_comparison}
\alias{validate_comparison}
\title{Validator for the 'comparison' class}
\usage{
validate_comparison(object)
}
\arguments{
\item{object}{A 'comparison' object.}
}
\value{
None. Errors when an invalid 'comparison' object is provided.
}
\description{
Validator for the 'comparison' class
}
