% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{validate_scenario}
\alias{validate_scenario}
\title{Validate a \code{scenario} class object}
\usage{
validate_scenario(object)
}
\arguments{
\item{object}{A \code{scenario} object}
}
\value{
None. Checks the validity of the provided \code{scenario} object.
}
\description{
Validate a \code{scenario} class object
}
