% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{comparison}
\alias{comparison}
\title{Create a \code{comparison} object}
\usage{
comparison(..., baseline, match_variables, comparison_variables)
}
\arguments{
\item{...}{Multiple \code{scenario}s or a list of \code{scenario} objects. At least one
of these scenarios, the 'baseline' scenario, must be named for the comparison
to be correctly constructed. The baseline scenario name must be the same as
passed to \code{baseline}.}

\item{baseline}{A string for the element of the list of \code{scenario}
objects which indicates which should be considered the 'baseline' outcome,
against which other outcomes are compared.}

\item{match_variables}{The variables in the \code{scenario} outputs on which to
match the scenarios and check whether they are comparable.}

\item{comparison_variables}{The variables in the \code{scenario} outputs to
compare against the 'baseline' scenario.}
}
\value{
A \code{comparison} object
}
\description{
The \code{comparison} class is intended to store \code{scenario}s and to
compare among them. One \code{scenario} must be set as the 'baseline' for such
comparisons.
}
\examples{
# prepare two scenarios of the final size of an epidemic
pandemic_flu <- scenario(
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(r0 = 1.5),
  replicates = 1L
)

covid19 <- scenario(
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(r0 = 5.0),
  replicates = 1L
)

# create a comparison object
comparison(
  pandemic_flu = pandemic_flu, covid19 = covid19,
  baseline = "pandemic_flu"
)

# pass scenario objects as a list
# create a comparison object
comparison(
  list(pandemic_flu = pandemic_flu, covid19 = covid19),
  baseline = "pandemic_flu"
)
}
