% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{comparison}
\alias{comparison}
\title{Create a 'comparison' object}
\usage{
comparison(..., baseline = 1L, match_variables, comparison_variables)
}
\arguments{
\item{...}{Multiple 'scenario's or a list of 'scenario' objects.}

\item{baseline}{A string or integer for the element of the list of 'scenario'
objects which indicates which should be considered the 'baseline' outcome,
against which other outcomes are compared. Taken to be the first element from
among the 'scenario' objects provided.}

\item{match_variables}{The variables in the 'scenario' outputs on which to
match the scenarios and check whether they are comparable.}

\item{comparison_variables}{The variables in the 'scenario' outputs to
compare against the 'baseline' scenario.}
}
\value{
A 'comparison' object
}
\description{
The 'comparison' class is intended to store 'scenario's and to
compare among them. One 'scenario' must be set as the 'baseline' for such
comparisons.
}
\examples{
# prepare two scenarios of the final size of an epidemic
pandemic_flu <- scenario(
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(r0 = 1.5),
  replicates = 1L
)

covid19 <- scenario(
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(r0 = 5.0),
  replicates = 1L
)

# create a comparison object
comparison(
  pandemic_flu = pandemic_flu, covid19 = covid19,
  baseline = "pandemic_flu"
)
}
