% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{new_comparison}
\alias{new_comparison}
\title{Constructor for the \code{comparisons} class}
\usage{
new_comparison(data, baseline = 1, matching_vars, comparison_vars)
}
\arguments{
\item{data}{A list of \code{scenario} objects.}

\item{baseline}{A string or integer for the element of the list of \code{scenario}
objects which indicates which should be considered the 'baseline' outcome,
against which other outcomes are compared.}

\item{matching_vars}{The variables in the \code{scenario} outputs on which to
match the scenarios and check whether they are comparable.}

\item{comparison_vars}{The variables in the \code{scenario} outputs to compare
against the 'baseline' scenario.}
}
\value{
A \code{comparison} object
}
\description{
Create a comparisons object after input checks.
}
\keyword{internal}
