% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{new_scenario}
\alias{new_scenario}
\title{Constructor for the \code{scenario} class}
\usage{
new_scenario(r0 = numeric(1), replicates = integer(1))
}
\arguments{
\item{r0}{The basic reproductive number \eqn{R_0} of the infection.}

\item{replicates}{The number of scenario replicates.}
}
\value{
scenario object
}
\description{
Create a scenario object check inputs.
}
\examples{
scenarios:::new_scenario(
 r0 = 1.5, replicates = 10
)
}
\keyword{internal}
