% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{new_scenario}
\alias{new_scenario}
\title{Constructor for the \code{scenario} class}
\usage{
new_scenario(
  name,
  model_function,
  parameters,
  extra_info = list(),
  replicates = integer(1)
)
}
\arguments{
\item{name}{The scenario name as a string. Defaults to \code{NA} if not provided.}

\item{model_function}{Function that is expected to run an epidemic scenario
model, such as \code{\link[finalsize:final_size]{finalsize::final_size()}}, as a string e.g.
"finalsize::final_size". Explicit namespacing is preferred.}

\item{parameters}{Parameters to the \code{model_function}.}

\item{extra_info}{Extra model information that may be useful for matching
models.}

\item{replicates}{The number of scenario replicates. This is the number of
times the model_function is run.}
}
\value{
A \code{scenario} object
}
\description{
Create a scenario object with input checks.
}
\keyword{internal}
