% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scenario.R
\name{run_scenario}
\alias{run_scenario}
\title{Run a scenario}
\usage{
run_scenario(x)
}
\arguments{
\item{x}{A \code{scenario} or \code{comparison} object.}
}
\value{
The original object with the \code{data} field populated with
simulation output. This object must be assigned.
Returns a \code{scenario} object when called on a \code{scenario} object as expected,
with the \code{data} field populated with epidemic simulation model outputs.
Returns a \code{comparison} object with the \code{data} fields of the \code{scenario}
objects contained in it populated.
}
\description{
Run an epidemic model scenario using the function stored in
\code{model_function} with the arguments in \code{parameters}. Runs as many replicates
of the function as specified in \code{replicates}. The simulation output data are
stored under \code{data}.
}
