% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sce_get_information}
\alias{sce_get_information}
\title{Get scenario information}
\usage{
sce_get_information(x, which)
}
\arguments{
\item{x}{A 'scenario' object.}

\item{which}{Which parameters to print.}
}
\value{
Nothing. Prints a list of parameters to screen.
}
\description{
Prints the value of an object in the \code{parameters} or in the
extra information list \code{extra_info}.
}
\examples{
# create a scenario
scenario_pandemic_flu <- scenario(
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(),
  replicates = 1
)

# get all parameters
sce_get_information(scenario_pandemic_flu)

# get only some parameters
sce_get_information(scenario_pandemic_flu, which = c("r0", "solver"))
}
