% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sce_get_information}
\alias{sce_get_information}
\title{Get scenario information}
\usage{
sce_get_information(x, which_info)
}
\arguments{
\item{x}{A \code{scenario} object.}

\item{which_info}{Which parameters or extra information to return.}
}
\value{
A named list with named elements corresponding to the names passed in
\code{which_info}.
}
\description{
Gets the value of one or more object in the \code{parameters} or the
extra information list \code{extra_info}.
}
