% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcomes.R
\name{sce_get_outcomes}
\alias{sce_get_outcomes}
\title{Get scenario outcomes}
\usage{
sce_get_outcomes(x)
}
\arguments{
\item{x}{A \code{scenario} or \code{comparison} object with the scenario outcome data
prepared. Check for whether data has been prepared using \code{\link[=sce_has_data]{sce_has_data()}}.}
}
\value{
A single \code{data.table} holding the output of all replicates of the
scenario. If \code{x} is a \code{comparison} object, the \code{data.table} holds the
outcomes of each replicate of each scenario specified therein. Scenarios in
this case are identified either by their names if any, or by a simple
synthetic identifier such as 'scenario_01'.
}
\description{
Get scenario outcomes
}
\examples{
# For the `scenario` class
# create a scenario specification
scenario_pandemic_flu <- scenario(
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(),
  replicates = 3 # note extra replicates
)

# run scenario
scenario_pandemic_flu <- run_scenario(scenario_pandemic_flu)

# get outcomes
sce_get_outcomes(scenario_pandemic_flu)

# For the `comparison` class
# prepare two scenarios of the final size of an epidemic
pandemic_flu <- scenario(
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(r0 = 1.5),
  replicates = 1L
)

covid19 <- scenario(
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(r0 = 5.0),
  replicates = 1L
)

# create a comparison object
outbreak_comparison <- comparison(
  pandemic_flu = pandemic_flu, covid19 = covid19,
  baseline = "pandemic_flu"
)

outbreak_comparison <- run_scenario(outbreak_comparison)

sce_get_outcomes(outbreak_comparison)

}
