% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sce_has_data}
\alias{sce_has_data}
\title{Check for scenario data}
\usage{
sce_has_data(x)
}
\arguments{
\item{x}{A \code{scenario} or \code{comparison} object.}
}
\value{
Whether the \code{scenario} has data, or whether all \code{scenario} objects in
a \code{comparison} object have data generated.
}
\description{
Check for scenario data
}
\examples{
# create a scenario
pandemic_flu <- scenario(
  name = "pandemic_flu",
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(),
  replicates = 1
)
covid19 <- scenario(
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(r0 = 3.0),
  replicates = 1
)

# for a `scenario` object
sce_has_data(pandemic_flu)

# for a `comparison` object
comparison_flu_covid <- comparison(
  pandemic_flu, covid19,
  baseline = "pandemic_flu"
)
sce_has_data(comparison_flu_covid)
}
