% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_helpers.R
\name{sce_set_baseline}
\alias{sce_set_baseline}
\title{Change the baseline scenario in a comparison}
\usage{
sce_set_baseline(x, new_baseline)
}
\arguments{
\item{x}{A \code{comparison} object.}

\item{new_baseline}{The name of a \code{scenario} object contained in the
\code{comparison} which is to be considered the new baseline.}
}
\value{
The \code{comparison} object with the baseline tag set to the new baseline
.
}
\description{
Change the baseline scenario in a comparison
}
\examples{
# create some scenarios
pandemic_flu <- scenario(
  name = "pandemic_flu",
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(r0 = 1.5),
  replicates = 1L
)

covid19 <- scenario(
  name = "covid19",
  model_function = "finalsize::final_size",
  parameters = make_parameters_finalsize_UK(r0 = 5.0),
  replicates = 1L
)

# create comparison, passing prepared scenarios with some names
outbreak_comparison <- comparison(
  pandemic_flu, covid19,
  baseline = "pandemic_flu"
)

# change baseline
sce_set_baseline(
  outbreak_comparison, "covid19"
)

}
