% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{scenario}
\alias{scenario}
\title{Create a \code{scenario} object}
\usage{
scenario(r0, replicates = 1L)
}
\arguments{
\item{r0}{The basic reproductive number \eqn{R_0} of the infection.}

\item{replicates}{The number of scenario replicates.}
}
\value{
A \code{scenario} object
}
\description{
The \code{scenario} class is intended to store the outcomes of a
number of runs of an epidemic simulation. This is a work in progress, and is
initially targeted for compatibility with outputs from
\code{\link[finalsize:final_size]{finalsize::final_size()}}.
}
\examples{
pandemic_influenza <- scenario(
  r0 = 1.5, replicates = 10
)
}
