% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios-package.R
\docType{package}
\name{scenarios-package}
\alias{scenarios}
\alias{scenarios-package}
\title{scenarios: Compare Epidemic Scenarios}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Compare epidemic scenarios such as those from the 'finalsize' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epiverse-trace/scenarios}
  \item \url{https://epiverse-trace.github.io/scenarios/}
  \item Report bugs at \url{https://github.com/epiverse-trace/scenarios/issues}
}

}
\author{
\strong{Maintainer}: Pratik Gupte \email{pratik.gupte@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5294-7819}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Rosalind Eggo \email{rosalind.eggo@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-0362-6717}{ORCID}) [copyright holder]
  \item Edwin Van Leeuwen \email{edwin.vanleeuwen@ukhsa.gov.uk} (\href{https://orcid.org/0000-0002-2383-5305}{ORCID}) [copyright holder]
}

}
\keyword{internal}
