\name{scenarios-package}
\alias{scenarios-package}
\alias{scenarios}

\docType{package}

\title{Scenario-based Stochastic Optimization}

\description{Scenario-based Stochastic Optimization}

\details{
\tabular{ll}{
Package: \tab scenarios\cr
Type: \tab Package\cr
Version: \tab 0.1-0\cr
Date: \tab 2014-09-25\cr
License: \tab GPL-3\cr
}
}

\author{
Ronald Hochreiter \email{ron@hochreiter.net}

Maintainer: Ronald Hochreiter \email{ron@hochreiter.net}
}

\references{\url{http://www.github.com/rhochreiter/scenarios/}}

\keyword{ package }

\examples{}