% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{validate_comparison}
\alias{validate_comparison}
\title{Validator for the \code{comparison} class}
\usage{
validate_comparison(object)
}
\arguments{
\item{object}{A \code{comparison} object.}
}
\value{
None. Errors when an invalid \code{comparison} object is provided.
}
\description{
Validator for the \code{comparison} class
}
