% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{validate_scenario}
\alias{validate_scenario}
\title{Validator for the \code{scenario} class}
\usage{
validate_scenario(scenario)
}
\arguments{
\item{scenario}{A \code{scenario} object}
}
\value{
None. Errors when an invalid \code{scenario} object is provided.
}
\description{
Validator for the \code{scenario} class
}
